backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.24.0
google-generativeai>=0.4.0
google-cloud-aiplatform>=1.42.0
google-cloud-sql-connector>=1.4.5
google-auth>=2.28.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
